split 'vocabulaire' module into store/actions/getters/mutations
